% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_split.R
\name{ingest_split}
\alias{ingest_split}
\title{Train/Test Split}
\usage{
ingest_split(
  train_test_raw_path = as.character(),
  holdout_raw_path = as.character(),
  target = as.character(),
  prop = as.numeric()
)
}
\arguments{
\item{train_test_raw_path}{Location of the raw train & test csv file to be ingested.}

\item{holdout_raw_path}{Location of the raw holdout csv file to be ingested.}

\item{target}{Name of the target column to be predicted in the datasets.}

\item{prop}{The proportion of train data to model.}
}
\value{


list of data.

train: The tibble for training the model.

test: The tibble for testing the model.

holdout: The tibble of unseen data to predict values for.
}
\description{
Ingest the training dataset and holdout datasets and split the training
dataset into train and test.
}
\examples{
\dontrun{
data_list = ingest_split(
     train_test_raw_path="path/to/ingest.csv",
     holdout_raw_path="path/to/export.csv",
     target="target_col",
     prop = 0.8
)
}
}
